Reduce number API calls, update version number

+++ app/src/main/java/stepn/sidekick/stepnsidekick/SidekickApi.java
@@ -6,7 +6,7 @@ import retrofit2.http.Path;
 
 public interface SidekickApi {
     @GET("mb/{energy}")
-    Call<MbChances> getMbChances(@Path("energy") double energy);
+    Call<MbChances> getMbChances(@Path("energy") float energy);
 
     @GET("gmt/numbers")
     Call<GmtMagicNumbers> getGmtNumbers();
